+++ be/src/vec/data_types/serde/data_type_quantilestate_serde.h
@@ -131,8 +131,8 @@ public:
             }
         }
     }
-    void read_column_from_arrow(IColumn& column, const arrow::Array* arrow_array, int start,
-                                int end, const cctz::time_zone& ctz) const override {
+    void read_column_from_arrow(IColumn& column, const arrow::Array* arrow_array, int64_t start,
+                                int64_t end, const cctz::time_zone& ctz) const override {
         throw doris::Exception(ErrorCode::NOT_IMPLEMENTED_ERROR,
                                "read_column_from_arrow with type " + column.get_name());
     }
