+++ be/src/vec/data_types/serde/data_type_struct_serde.h
@@ -18,13 +18,13 @@
 #pragma once
 
 #include <glog/logging.h>
-#include <stdint.h>
 
-#include <ostream>
+#include <cstdint>
 
 #include "common/status.h"
 #include "data_type_serde.h"
 #include "util/jsonb_writer.h"
+#include "vec/io/reader_buffer.h"
 
 namespace doris {
 class PValues;
@@ -150,8 +150,8 @@ public:
     void write_column_to_arrow(const IColumn& column, const NullMap* null_map,
                                arrow::ArrayBuilder* array_builder, int64_t start, int64_t end,
                                const cctz::time_zone& ctz) const override;
-    void read_column_from_arrow(IColumn& column, const arrow::Array* arrow_array, int start,
-                                int end, const cctz::time_zone& ctz) const override;
+    void read_column_from_arrow(IColumn& column, const arrow::Array* arrow_array, int64_t start,
+                                int64_t end, const cctz::time_zone& ctz) const override;
 
     Status write_column_to_mysql(const IColumn& column, MysqlRowBuffer<true>& row_buffer,
                                  int64_t row_idx, bool col_const,
